# apps/school/src/data/swagger.json
{"openapi":"3.0.0","paths":{"/organization":{"post":{"operationId":"OrganizationController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrganizationDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"OrganizationController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/organization/{id}":{"get":{"operationId":"OrganizationController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"OrganizationController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrganizationDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"OrganizationController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/address":{"post":{"operationId":"AddressController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAddressDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"AddressController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/address/{id}":{"get":{"operationId":"AddressController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"AddressController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAddressDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"AddressController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/classes":{"post":{"operationId":"ClassesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClassesDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"ClassesController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/classes/by-grade/{gradeId}":{"get":{"operationId":"ClassesController_fetchByGrade","parameters":[{"name":"gradeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/classes/{id}":{"get":{"operationId":"ClassesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"ClassesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateClassesDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"ClassesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/grades":{"post":{"operationId":"GradesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGradeDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"GradesController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/grades/with-classes":{"get":{"operationId":"GradesController_findAllWithClasses","parameters":[],"responses":{"200":{"description":""}}}},"/grades/{id}":{"get":{"operationId":"GradesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"GradesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGradeDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"GradesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/subject":{"post":{"operationId":"SubjectController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubjectDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"SubjectController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/subject/{id}":{"get":{"operationId":"SubjectController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"SubjectController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSubjectDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"SubjectController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/class-subjects":{"post":{"operationId":"ClassSubjectsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClassSubjectsDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"ClassSubjectsController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/class-subjects/{id}":{"get":{"operationId":"ClassSubjectsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"ClassSubjectsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateClassSubjectDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"ClassSubjectsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/schools":{"post":{"operationId":"creates school along with organization and address","parameters":[{"name":"user-agent","in":"header","description":"User agent header","required":false,"schema":{"default":"Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10.7.6; rv:7.6) Gecko/20100101 Firefox/7.6.5","example":"Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10.7.6; rv:7.6) Gecko/20100101 Firefox/7.6.5","type":"string"}},{"name":"x-timestamp","in":"header","description":"Timestamp header, in microseconds","required":false,"schema":{"default":1697804202329,"example":1697804202329,"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchoolCreateDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"school.create"},"statusCode":{"type":"number","example":201}}}}}},"408":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"http.serverError.requestTimeout"},"statusCode":{"type":"number","example":5052}}}}}},"422":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"request.validation"},"statusCode":{"type":"number","example":5070}}}}}},"500":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"http.serverError.internalServerError"},"statusCode":{"type":"number","example":5050}}}}}},"503":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"http.serverError.serviceUnavailable"},"statusCode":{"type":"number","example":5051}}}}}}}},"get":{"operationId":"SchoolsController_findAll","summary":"get all of school and its detail of organization and address","parameters":[{"name":"user-agent","in":"header","description":"User agent header","required":false,"schema":{"default":"Opera/12.19 (Windows NT 5.0; U; BN Presto/2.9.175 Version/11.00)","example":"Opera/12.19 (Windows NT 5.0; U; BN Presto/2.9.175 Version/11.00)","type":"string"}},{"name":"x-timestamp","in":"header","description":"Timestamp header, in microseconds","required":false,"schema":{"default":1697804202332,"example":1697804202332,"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"school.get"},"statusCode":{"type":"number","example":200},"data":{"$ref":"#/components/schemas/SchoolListSerialization"}}}}}},"408":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"http.serverError.requestTimeout"},"statusCode":{"type":"number","example":5052}}}}}},"500":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"http.serverError.internalServerError"},"statusCode":{"type":"number","example":5050}}}}}},"503":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"http.serverError.serviceUnavailable"},"statusCode":{"type":"number","example":5051}}}}}}}}},"/schools/{id}":{"get":{"operationId":"SchoolsController_findOne","summary":"get detail of single school along with organization and address","parameters":[{"name":"id","required":true,"in":"path","allowEmptyValue":false,"example":"7d0a9b58-11b7-4ec2-854f-b6f997906709","schema":{"type":"string"}},{"name":"user-agent","in":"header","description":"User agent header","required":false,"schema":{"default":"Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10_5_1 rv:6.0; FO) AppleWebKit/532.1.2 (KHTML, like Gecko) Version/5.0.3 Safari/532.1.2","example":"Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10_5_1 rv:6.0; FO) AppleWebKit/532.1.2 (KHTML, like Gecko) Version/5.0.3 Safari/532.1.2","type":"string"}},{"name":"x-timestamp","in":"header","description":"Timestamp header, in microseconds","required":false,"schema":{"default":1697804202333,"example":1697804202333,"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"school.get"},"statusCode":{"type":"number","example":200},"data":{"$ref":"#/components/schemas/SchoolGetSerialization"}}}}}},"408":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"http.serverError.requestTimeout"},"statusCode":{"type":"number","example":5052}}}}}},"500":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"http.serverError.internalServerError"},"statusCode":{"type":"number","example":5050}}}}}},"503":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"http.serverError.serviceUnavailable"},"statusCode":{"type":"number","example":5051}}}}}}}},"patch":{"operationId":"SchoolsController_update","summary":"update school along with organization and address","parameters":[{"name":"id","required":true,"in":"path","allowEmptyValue":false,"example":"7d0a9b58-11b7-4ec2-854f-b6f997906709","schema":{"type":"string"}},{"name":"user-agent","in":"header","description":"User agent header","required":false,"schema":{"default":"Mozilla/5.0 (compatible; MSIE 8.0; Windows NT 6.1; Trident/5.1)","example":"Mozilla/5.0 (compatible; MSIE 8.0; Windows NT 6.1; Trident/5.1)","type":"string"}},{"name":"x-timestamp","in":"header","description":"Timestamp header, in microseconds","required":false,"schema":{"default":1697804202333,"example":1697804202333,"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchoolUpdateDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"school.update"},"statusCode":{"type":"number","example":200},"data":{"$ref":"#/components/schemas/SchoolGetSerialization"}}}}}},"408":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"http.serverError.requestTimeout"},"statusCode":{"type":"number","example":5052}}}}}},"422":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"request.validation"},"statusCode":{"type":"number","example":5070}}}}}},"500":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"http.serverError.internalServerError"},"statusCode":{"type":"number","example":5050}}}}}},"503":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDefaultSerialization"}],"properties":{"message":{"example":"http.serverError.serviceUnavailable"},"statusCode":{"type":"number","example":5051}}}}}}}},"delete":{"operationId":"SchoolsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"School APIs Specification","description":"Section for describe whole APIs","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"/"}],"components":{"schemas":{"CreateOrganizationDto":{"type":"object","properties":{"name":{"type":"string","example":"Example Organization Ltd."},"details":{"type":"object","example":{"website":"https://example.org","address":"123 Main St","contact":"contact@example.org"}}},"required":["name","details"]},"UpdateOrganizationDto":{"type":"object","properties":{}},"CreateAddressDto":{"type":"object","properties":{"street":{"type":"string","example":"123 Main Street"},"street1":{"type":"string","example":"Apt 4B"},"council":{"type":"string","example":"Council Name"},"town":{"type":"string","example":"Townsville"},"tehsil":{"type":"string","example":"Tehsil Area"},"district":{"type":"string","example":"Districtville"},"county":{"type":"string","example":"County County"},"state":{"type":"string","example":"State Province"},"country":{"type":"string","example":"Country Nation"},"latitude":{"type":"number","example":42.123456},"longitude":{"type":"number","example":-71.654321},"addressDetails":{"type":"object","example":{"key1":"value1","key2":"value2"}}},"required":["street","street1","council","town","tehsil","district","county","state","country","latitude","longitude","addressDetails"]},"UpdateAddressDto":{"type":"object","properties":{}},"CreateClassesDto":{"type":"object","properties":{}},"UpdateClassesDto":{"type":"object","properties":{}},"CreateGradeDto":{"type":"object","properties":{}},"UpdateGradeDto":{"type":"object","properties":{}},"CreateSubjectDto":{"type":"object","properties":{}},"UpdateSubjectDto":{"type":"object","properties":{}},"CreateClassSubjectsDto":{"type":"object","properties":{}},"UpdateClassSubjectDto":{"type":"object","properties":{}},"ResponseDefaultSerialization":{"type":"object","properties":{"statusCode":{"type":"number","nullable":false,"description":"return specific status code for every endpoints","example":200},"message":{"nullable":false,"description":"Message base on language","oneOf":[{"type":"string","example":"message endpoint"},{"type":"object","example":{"en":"This is test endpoint.","id":"Ini adalah endpoint test"}}]},"_metadata":{"type":"object","nullable":false,"description":"Contain metadata about API","example":{"timestamp":1660190937231,"timezone":"Asia/Karachi","requestId":"40c2f734-7247-472b-bc26-8eff6e669781","path":"/api/v1/test/hello","version":"1","repoVersion":"1.0.0"}}},"required":["statusCode","message","_metadata"]},"SchoolCreateDto":{"type":"object","properties":{"name":{"type":"string","example":"Kling and Sons"},"model":{"type":"string","example":"deserunt"},"status":{"type":"string","example":"error"},"offline":{"type":"number","example":1},"started":{"format":"date-time","type":"string","example":"2023-07-30T09:17:50.531Z"},"electricity":{"type":"string","example":"amet"},"internet":{"type":"string","example":"https://classic-ravioli.biz/"},"schoolType":{"type":"string","example":"velit"},"estimatedHousehold":{"type":"number","example":1802934994075648},"spaceType":{"type":"string","example":"maiores"},"nearestSchool":{"type":"string","example":"sequi"},"details":{"type":"object","example":{"anyField":"anyValue"}},"address":{"example":{"street":"123 Main Street","street1":"Apt 4B","council":"Council Name","town":"Town Name","tehsil":"Tehsil Name","district":"District Name","county":"County Name","state":"State Name","country":"Country Name","latitude":42.123456,"longitude":-71.654321,"addressDetails":{"detail1":"value1","detail2":"value2"}},"allOf":[{"$ref":"#/components/schemas/CreateAddressDto"}]},"organization":{"example":{"name":"Example Organization Name","details":{"key1":"value1","key2":"value2"}},"allOf":[{"$ref":"#/components/schemas/CreateOrganizationDto"}]}},"required":["name","model","status","offline","started","electricity","internet","schoolType","estimatedHousehold","spaceType","nearestSchool","details","address","organization"]},"SchoolListSerialization":{"type":"object","properties":{"data":{"nullable":false,"description":"Object of all schools populated with organization and address","type":"array","items":{"type":"array"}}},"required":["data"]},"OrganizationGetSerialization":{"type":"object","properties":{"_id":{"type":"string","description":"Id that representative with your target data","example":"d73b084f-b6d1-4e86-bd9e-8c1ef46d390d","nullable":false},"updatedBy":{"type":"string","description":"Document updated by if [probably admin]","example":"admin_user_id","nullable":false},"effectiveAt":{"format":"date-time","type":"string","description":"Date for document to be effective at","example":"2023-10-20T12:16:42.328Z","nullable":false},"createdAt":{"format":"date-time","type":"string","description":"Date created at","example":"2023-10-20T12:16:42.328Z","nullable":false},"updatedAt":{"format":"date-time","type":"string","description":"Date updated at","example":"2023-10-20T12:16:42.328Z","nullable":false},"name":{"type":"string","example":"Example Organization Ltd."},"details":{"type":"object","example":{"website":"https://example.org","address":"123 Main St","contact":"contact@example.org"}}},"required":["_id","updatedBy","effectiveAt","createdAt","updatedAt","name","details"]},"AddressGetSerialization":{"type":"object","properties":{"_id":{"type":"string","description":"Id that representative with your target data","example":"d73b084f-b6d1-4e86-bd9e-8c1ef46d390d","nullable":false},"updatedBy":{"type":"string","description":"Document updated by if [probably admin]","example":"admin_user_id","nullable":false},"effectiveAt":{"format":"date-time","type":"string","description":"Date for document to be effective at","example":"2023-10-20T12:16:42.328Z","nullable":false},"createdAt":{"format":"date-time","type":"string","description":"Date created at","example":"2023-10-20T12:16:42.328Z","nullable":false},"updatedAt":{"format":"date-time","type":"string","description":"Date updated at","example":"2023-10-20T12:16:42.328Z","nullable":false},"street":{"type":"string","example":"123 Main Street"},"street1":{"type":"string","example":"Apt 4B"},"council":{"type":"string","example":"Council Name"},"town":{"type":"string","example":"Townsville"},"tehsil":{"type":"string","example":"Tehsil Area"},"district":{"type":"string","example":"Districtville"},"county":{"type":"string","example":"County County"},"state":{"type":"string","example":"State Province"},"country":{"type":"string","example":"Country Nation"},"latitude":{"type":"number","example":42.123456},"longitude":{"type":"number","example":-71.654321},"addressDetails":{"type":"object","example":{"key1":"value1","key2":"value2"}}},"required":["_id","updatedBy","effectiveAt","createdAt","updatedAt","street","street1","council","town","tehsil","district","county","state","country","latitude","longitude","addressDetails"]},"SchoolGetSerialization":{"type":"object","properties":{"_id":{"type":"string","description":"Id that representative with your target data","example":"d73b084f-b6d1-4e86-bd9e-8c1ef46d390d","nullable":false},"updatedBy":{"type":"string","description":"Document updated by if [probably admin]","example":"admin_user_id","nullable":false},"effectiveAt":{"format":"date-time","type":"string","description":"Date for document to be effective at","example":"2023-10-20T12:16:42.328Z","nullable":false},"createdAt":{"format":"date-time","type":"string","description":"Date created at","example":"2023-10-20T12:16:42.328Z","nullable":false},"updatedAt":{"format":"date-time","type":"string","description":"Date updated at","example":"2023-10-20T12:16:42.328Z","nullable":false},"name":{"type":"string","example":"Grimes - Wisoky"},"model":{"type":"string","example":"ut"},"status":{"type":"string","example":"sequi"},"offline":{"type":"number","example":1},"started":{"format":"date-time","type":"string","example":"2023-06-06T11:47:07.930Z"},"electricity":{"type":"string","example":"cupiditate"},"internet":{"type":"string","example":"https://naive-rubber.com"},"schoolType":{"type":"string","example":"voluptates"},"estimatedHousehold":{"type":"number","example":642567909670912},"spaceType":{"type":"string","example":"consequatur"},"nearestSchool":{"type":"string","example":"nobis"},"details":{"type":"object","example":{"anyField":"anyValue"}},"organization":{"nullable":false,"default":[],"allOf":[{"$ref":"#/components/schemas/OrganizationGetSerialization"}]},"address":{"nullable":false,"default":[],"allOf":[{"$ref":"#/components/schemas/AddressGetSerialization"}]}},"required":["_id","updatedBy","effectiveAt","createdAt","updatedAt","name","model","status","offline","started","electricity","internet","schoolType","estimatedHousehold","spaceType","nearestSchool","details","organization","address"]},"SchoolUpdateDto":{"type":"object","properties":{"address":{"example":{"street":"123 Main Street","street1":"Apt 4B","council":"Council Name","town":"Town Name","tehsil":"Tehsil Name","district":"District Name","county":"County Name","state":"State Name","country":"Country Name","latitude":42.123456,"longitude":-71.654321,"addressDetails":{"detail1":"value1","detail2":"value2"}},"allOf":[{"$ref":"#/components/schemas/CreateAddressDto"}]},"organization":{"example":{"name":"Example Organization Name","details":{"key1":"value1","key2":"value2"}},"allOf":[{"$ref":"#/components/schemas/CreateOrganizationDto"}]}}}}}}